sbom-utility-scripts/scripts/sbom-for-modelcar-task/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
packageurl-python==0.16.0
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
